Correct quiz deleting for correct importing in future

+++ com_joomlaquiz_expert/admin/models/quizzes.php
@@ -310,6 +310,25 @@ class JoomlaquizModelQuizzes extends JModelList
 		$db = JFactory::getDBO();
 		$option = "com_joomlaquiz";
 		if (count( $cid )) {
+			
+			$query = "DELETE FROM #__quiz_pool" . "\n WHERE " . $db->qn('q_id') . " IN ( ". implode(', ', $cid) . ")";
+            $db->setQuery( $query );
+            $db->execute();
+
+            $query = "DELETE FROM #__quiz_feed_option" . "\n WHERE " . $db->qn('quiz_id') . " IN ( ". implode(', ', $cid) . ")";
+            $db->setQuery( $query );
+            $db->execute();
+
+			$names = "'com_joomlaquiz.quiz.".implode( "', 'com_joomlaquiz.quiz.", $cid )."'";
+            $query = "DELETE FROM #__assets"
+                . "\n WHERE ".$db->qn('name')." IN ( $names )"
+            ;
+            $db->setQuery( $query );
+            if (!$db->execute()) {
+                echo "<script> alert('".$db->getErrorMsg()."'); window.history.go(-1); </script>\n";
+                exit();
+            }
+
 			$cids = implode( ',', $cid );
 			$query = "DELETE FROM #__quiz_t_quiz"
 			. "\n WHERE c_id IN ( $cids )"
